Helping to address issue #5
Issue #5 is hard to test, but one thing that I was able to repro is to
find a channel that was offline AND started re-broadcasting another one.

In this scenario, I noticed that when the re-broadcasted channel was up,
the extension would fail to inject itself.

This commits fixes the failure to inject by fetching the video player div
through its id (duh) instead of using a class name, which was different
for a re-broadcasted stream.