Fix re-attaching plugin when navigating Twitch
Twitch is a Single Page App. They do some nice magic when clicking on
some of the links, like for example when watching a specific channel and
then clicking on a link with the name of the game being streamed.

Because it's navigating to a page where there is no video player, the
bullet text extension disables itself (and shows a yellow icon instead
of a green one). It will stay like that until the user navigates back
to a page with a gamer streaming a game, which gets detected and triggers
the bullet text extension to re-enable itself (and shows a green icon
again on the url bar). Magic.